feat: Add `home_dir` param to `codex_cli` agent
Allows the home directory of Codex to be configured. This is the directory where the global `AGENTS.md` and MCP configuration files are read from, and where session logs are written to.

Configuring Codex home allows these files not to clutter up the main workspace or get included in any workspace diffs done during an eval.

This commit does not change the default, which is the sandbox working directory, so is backward compatible.